Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > Attempted a butchered rewording of the commit comment.
 > 
 > I'm still a little unclear on the motivation for the original
 > check, which I tried to capture in the description. Also, didn't
 > TLS 1.3 make it fatal to ignore the SignatureAndHashAlgorithm in
 > the SKE? I haven't caught up on the whole new key derivation
 > scheme, so it may be entirely outdated for me to even mention that.

TLS 1.2 makes the extension optional and you're supposed to interpret missing as {*, sha1}. Yeah, I think 1.3 is going to change that, which we'll certainly need to enforce once it comes time for that. For now, we're just a 1.2 implementation.

But that SHA-1 fallback is in the other direction, so it's not related to this function. (There is another SHA-1 fallback in this direction, which https://boringssl-review.googlesource.com/#/c/6411/ removes, because it's dumb.)

 > I'm fully inclined to let the caller (aka cert verifier) be
 > responsible for all cryptographic policies and criteria. e.g. if it
 > says "I want SHA-256" and the server sends SHA-1, the client should
 > decide whether to yell or not. The concern I would have, though, is
 > whether that means there's a security check that wouldn't happen -
 > for example, if the caller sets a SigAndHashAlgs to disable SHA-1,
 > the peer does a sig exchange with SHA-1, and the caller, not
 > knowing the need to check, inadvertently picks up SHA-1. The
 > existing code handles that (well, with the gross fallback to
 > SHA-1), so it's unclear where the right line to draw for the API
 > is.
 > 
 > That's the long-winded way of saying "I'm not sure; more info
 > needed"

So, to be clear, I have no intention of removing the sigalgs half of this check. It's just the curve check in this function that makes no sense as-is. (Whether we want to remove it or move it elsewhere is the question.)

It's true that there'll be some amount of enforcement in the TLS stack regardless because only the TLS stack can enforce that SKX signature hashes we accept. There actually isn't an API for the caller to configure that right now, but I expect we'll add one eventually when we take SHA-1 out of sigalgs[1]. Have a CL for the API even: https://boringssl-review.googlesource.com/#/c/6130/

Though this does mean that, no matter what, we can't punt all policy to the verifier. Hrm. I think I'm starting to lean more towards putting the curve check in the TLS stack. I dunno. I have no love for how all the random checks are just sprinkled haphazardly all over the place in OpenSSL, but I should be able to clean that up independently.

[1] Note: we can't remove that until BOTH SHA-1 certificates are gone and we've confirmed that SHA-1 SKX signatures don't exist. That's what https://boringssl-review.googlesource.com/#/c/6413/ is intended to measure. In so far as one can cleanly separate the two places to enforce policy, TLS doesn't do a very good job of letting the protocol advertise that.

Patch-set: 1
Label: Code-Review=0