Update patch set 1
Patch Set 1: Code-Review+2

Another option that I had pondered for this issue was to have the IANA database parsed into code somewhere. The advantage of that is that it would cover ciphersuites that we don't implement but the disadvantage is that dealing with cipher suites in development (like ChaCha20-Poly1305) would be more difficult.

This path seems like less code so let's go with this.

Patch-set: 1
Label: Code-Review=+2