Update patch set 7
Abandoned

While there probably is some merit to doing this, I don't think this is the right way to do it I'm afraid. It also makes the conditions in the code harder to follow. I don't feel strongly but I'm afraid that I'm landing against this change.

Patch-set: 7
Status: abandoned