Update patch set 6
Patch Set 6:

`ec_GFp_simple_oct2point` contains this code:

```
  if (form == POINT_CONVERSION_COMPRESSED) {
    if (!EC_POINT_set_compressed_coordinates_GFp(group, point, x, y_bit, ctx)) {
      goto err;
    }
  } else {
    if (!BN_bin2bn(buf + 1 + field_len, field_len, y)) {
      goto err;
    }
    if (BN_ucmp(y, &group->field) >= 0) {
      OPENSSL_PUT_ERROR(EC, EC_R_INVALID_ENCODING);
      goto err;
    }

    if (!EC_POINT_set_affine_coordinates_GFp(group, point, x, y, ctx)) {
      goto err;
    }
  }

  /* test required by X9.62 */
  if (!EC_POINT_is_on_curve(group, point, ctx)) {
    OPENSSL_PUT_ERROR(EC, EC_R_POINT_IS_NOT_ON_CURVE);
    goto err;
  }
```

In the case of a compressed point, doesn't the decompression ensure that the point is on the curve? If so, the explicit check that the point is on the curve isn't necessary for compressed points.

In the case of an uncompressed point, the call to `EC_POINT_set_affine_coordinates_GFp` ensures that the point is already on the curve, so the explicit check is not necessary either.

In the review comments above, it was noted that we want to try to avoid checking that the point is on the curve more than once. I think that is a good idea, but the current code doesn't achieve that because `ec_GFp_simple_oct2point` checks the point is on the curve twice. I believe the solution is to remove the explicit check in in `ec_GFp_simple_oct2point` in order to avoid the (small) performance regression in basically everything that uses ECC, including in particular crypto/x509's ECDSA support and libssl's ECDH and ECDSA support. It may be a good idea to add a benchmark to `bssl speed` to measure oct2point performance.

Note that `ec_GFp_simple_oct2point` also goes out of its way to allow an attacker to send you an encoded point at infinity. You probably should just get rid of that. Otherwise, that is something that also needs to be taken into consideration.

Patch-set: 6
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0