Update patch set 3
Patch Set 3:

> This change has been hanging around for a long time. Did you toy
 > with any variants? Frankly, I only half understand what's going on
 > here.

Nah, I think I left it in my queue forgot about it. :-) I think my commit message wasn't very good. I'll rewrite it if you think it's worth bothering with.

So, say you have some 128-byte buffer and all 128 bytes is the record. So the rule is that your output buffer, if it aliases, must be behind in + ssl_record_prefix_len. That's the record header and the length of any explicit nonce.

As tls_record_open is currently written, there's no alias check needed. The only time you write to out is SSL_AEAD_CTX_open and we assume SSL_AEAD_CTX_open internally knows to enforce that it won't any part of its input that it hasn't read yet.

However, consider if the input is a 256-byte buffer and the first 128 bytes is the record. The input we pass into SSL_AEAD_CTX is [in + 5, in+128). For sanity, let's say there is no explicit nonce. Then SSL_AEAD_CTX believes out is allowed to start before in + 5 or after in + 128. But that's not sufficient. I might call SSL_AEAD_CTX_open with out at in + 200, and then I'll clobber the rest of the input buffer.

I think the alternative would be to stick at the front of the function something like:

  if (out > in + ssl_record_prefix_len(ssl) && in + in_len <= out) {
    return 0;
  }

Although this too is a little confusing because pointer arithmetic is undefined when it escapes out of a buffer. Which means we actually need:

  if (in_len >= ssl_record_prefix_len(ssl) /* Avoid undefined behavior. */ &&
     out > in + ssl_record_prefix_len(ssl) && in + in_len <= out) {
    return ssl_record_open_error;
  }

Or probably more sensible:

  if (in_len < ssl_record_prefix_len(ssl)) {
    *out_consumed = ssl_record_prefix_len(ssl);
    return ssl_record_open_partial;
  }
  if (out > in + ssl_record_prefix_len(ssl) && in + in_len <= out) {
    return 0;
  }

This also doesn't capture the "discard a single record in a DTLS packet" case. Which we don't care about, so whatever.

Patch-set: 3