Update patch set 1
Patch Set 1:

It may be the case that only one of the two implementations is necessary. In theory the binary inverse algorithm should be faster for smallish N, but it is probably worth verifying at some point that this implementation of it actually is faster in some important situation (platform * algorithm). And/or it is worth considering whether the code needs to support even N, or whether it can restrict itself to odd N only so that the binary inversion can always be used.

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0