Update patch set 1
Patch Set 1:

(1 comment)

I looked at this patch and it looks good to me. (I was already working on a very similar patch.)

The main question I have is whether it is a good idea to use the same names for the now-unsigned values that were used for signed values. Looking at the code, it looks like nothing will go wrong, but it's hard to say whether upstream will ever do something non-obvious that might depend on these values being sign-extended or whatnot. Also, the existing names are not very good. I think it would be good to mint new names that are more descriptive and that don't have any possibility of conflicting with upstream usage.

Patch-set: 1
Label: Code-Review=0