Fold EC_GROUP_new_curve_GFp and EC_GROUP_set_generator into a EC_GROUP_new_arbitrary.
This is only for Conscrypt which always calls the pair in succession. (Indeed
it wouldn't make any sense to not call it.) Remove those two APIs and replace
with a singled merged API. This way incomplete EC_GROUPs never escape outside
our API boundary and EC_GROUPs may *finally* be made immutable.

Add a temporary BORINGSSL_201512 define to ease the transition for Conscrypt.

Change-Id: I3706c2ceac31ed2313175ba5ee724bd5c74ef6e1