Improve the safety of point multiplication.
Split |EC_METHOD::mul| into two methods, one for multiplication by
secret scalars and one for multiplication by public scalars. Then,
move the imperfect mitigation for timing attacks against the
wNAF-based multiplication from the ECDSA signing code to
|ec_wNAF_mul_private| so that it doesn't negatively affect the
"constant-time" implementations. Then, remove the
mitigation-for-the-mitigation uses of |BN_nmod| from the
"constant-time" implementations, which made the "constant-time"
implementations not really constant-time.

More importantly, it is not clear if was OK for the "constant-time"
implementations to do the reduction only when
|BN_num_bits(X) > 256|. In particular, that condition meant that a
scalar X such that |group->order| <= X < 2^256 would not be reduced,
and it is not obvious--especially in the case of p256-x86_64--whether
that is OK.

It was not clear what to do about |EC_POINT_mul|, so it was made to be
strict about enforcing that its input scalars are in the range
[0, group->order - 1]|. If this breaks external users of that function,
then |EC_POINT_mul| could be change to instead use |BN_nmod| to reduce
its inputs prior to calling |group->meth->mul_private|.

Change-Id: Ib547b617b07b3e7000eab4b95b79cc011e5f8e5c