Let generate_build_files.py work with other directory hierarchies.
This change provides 3 new options:

  --src SRC   path to source directory
  --dst DST   path to destination directory
  --relative  write file paths relative to the source directory

which make this script useable in cases where BoringSSL's source code
isn't located in src/ directory.

This is especially useful when working with Bazel, which can use this
script to generate ASM files on-the-fly and store them in $GENDIR.

Defaults are preserved, so this script behaves the same as before
if no new options are specified.

Change-Id: I56115fad49fbd6c9bc51124624e2a390ec64ec95
Signed-off-by: Piotr Sikora <piotrsikora@google.com>