Avoid potential uninitialized memory read in p256-x86_64.c.
|ecp_nistz256_points_mul| sets the |top| of the result point's X, Y,
and Z coordinates to |P256_LIMBS|, then does the calculation, then
fills in their values. However, if an error occurs during the
calculation, the values are never filled in. If a caller were to then
attempt to read the result point's coordinates, then an uninitialized
memory read would occur. This is unlikely to happen in practice because
the caller shouldn't be trying to use the result point when the
function fails.

Change-Id: I8c48deb63163a27f739c8797962414f8ca2588cd