Update patch set 3
Patch Set 3:

I am pretty sure that using fprintf on a closed file descriptor is undefined behavior, and it may be the case that an application has already closed stderr. For example, daemons often do that. So it seems like a bad idea to have the |fprintf| in here.

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0