Simplify RSA key exchange padding check.
In the RSA key exchange, the expected size of the output is known, which makes
the padding check much simpler. Moreover, any context which could possibly hope
to handle RSAES-PKCS1-v1_5 without side channels must already know the length
of the output. (Otherwise any integrity check or swap to randomness or other
mitigation is basically doomed to fail.) So there isn't actually any use in
exporting the generic one.

Instead, replace it with a variant which knows the actual size and swaps in
randomness. In doing so, switch the handling of client_version from a
constant-time check to unconditionally setting the first two bytes to
client_version. This aligns with the recommendations in RFC 5246 and
http://eprint.iacr.org/2003/052.pdf and is slightly simpler. It also allows
rsa.h to expose an API with fewer sharp edges.

Verified with the valgrind uninitialized memory trick that we're still
constant-time.

Also update rsa.h to recommend against using the PKCS#1 v1.5 schemes.

Thanks to Ryan Sleevi for the suggestion to take advantange of the known
length.

Change-Id: I3f8950df4b2298074e33335bea4324fe65b83806