Update patch set 3
Abandoned

Abandoning this version. I think the extremely unlikely TODO(davidben) note in the commit message is also silly, but I'm pretty convinced at the point that the reasoning in the paper to prefer this version check is nonsense.

Of the two weaknesses, the second is easy to disregard. The "attack is":

Suppose the attacker knows all but the first two bytes of some 48-byte plaintext and has a PKCS#1-encrypted ciphertext of it. Then they could vary the version in the ClientHello, try all 256^2 possibilities, and find the one whose Finished message is accepted.

If, instead, the attacker knew all but the last two bytes of the plaintext, they could make the version in the ClientHello match the known first two bytes and try all 256^2 possibilities for the last two bytes and find the one whose Finished message is accepted. This works against both protocols.

The first weakness supposes the attacker can, via power or electromagnetic side channels, distinguish P being replaced with a random premaster from P being used deterministicly by seeing whether signals are sufficiently correlated with each other. But if that were possible, then the whole idea of replacing the premaster with randomness and suspect and one could use a PKCS#1 padding oracle rather than this version-check-based distinguisher.

Patch-set: 3
Status: abandoned