Update patch set 4
Patch Set 4:

> Thanks for the fast reply. Making the numbers opaque to the SSL
 > stack sounds fine to me, should I look into something for this?

If you like, sure! Not hugely high-priority since I imagine you won't need to add a new one for a while. But I don't believe we have any consumers of DTLS-SRTP other than WebRTC, so whatever API works best for WebRTC. I don't think we necessarily even have to keep the old API around if you don't need it.

(Hrm. Although if WebRTC still builds against OpenSSL, it might be inconvenient to have a different API. Meh. I defer to you all how you want to do this. Keeping things as-is is fine too.)

 > Also thanks for removing the comment, makes sense.
 > 
 > Anything else to do for this CL? (Sorry, I'm not very used to
 > Gerrit).

Nah. I'll just go ahead and merge this and roll it into Chromium since it's completely trivial and you're presumably blocked on it. (Adam's out for a bit. Normally I leave it to him to land things and give a final stamp.)

Patch-set: 4