Update patch set 3
Patch Set 3:

> I think this is a good change. But, wouldn't it be better to make
 > these real (inline) functions, especially since they evaluate their
 > arguments multiple times? Perhaps an OPENSSL_ALWAYS_INLINE that is
 > defined to __attribute(always_inline)__ for GCC/clang or
 > __forceinline for MSVC would be useful to maximize the inlining of
 > it, if that is the concern.
 > 
 > Note that these functions are used most heavily in OPENSSL_NO_ASM
 > builds, so it might be useful to do your measurements in that kind
 > of build.

Yeah, something of that sort was going to be the next change, and then I got distracted. :-) It actually needs to get split into two sets, depending on DATA_ORDER, which is necessary anyway to make m32_common.h less confused about whether it's a multiply-included or singly-included header.

Patch-set: 3