Implement draft-ietf-tls-curve25519-01 in C.
TODO(davidben): Take another pass over this and make sure it's good.

The new curve is advertised by default and placed at the front of the
list.

TODO(davidben): Should this be off by default initially?

As EC_GROUP/EC_POINT is a bit too complex for X25519, this introduces an
SSL_ECDH_METHOD abstraction which wraps just the raw ECDH operation. It
also tidies up some of the curve code which kept converting back and
force between NIDs and curve IDs. Now everything transits as curve IDs
except for API entry points (SSL_set1_curves) which take NIDs. Those
convert immediately and act on curve IDs from then on.

Note that, like the Go implementation, this slightly tweaks the order of
operations. The client sees the server public key before sending its
own. To keep the abstraction simple, SSL_ECDH_METHOD expects to
generate a keypair before consuming the peer's public key. Instead, the
client handshake stashes the serialized peer public value and defers
parsing it until it comes time to send ClientKeyExchange. (This is
analogous to what it was doing before where it stashed the parsed peer
public value instead.)

It still uses TLS 1.2 terminology everywhere, but this abstraction
should also be compatible with TLS 1.3 which unifies ECDHE and DHE (not
that there's much use in implementing the latter). I didn't rename
everything just yet because calling it a "key exchange" is confusing
when, in TLS 1.2, this also encompasses PSK, RSA, and DHE_*.

BUG=571231

Change-Id: Iba09dddee5bcdfeb2b70185308e8ab0632717932