Update patch set 2
Patch Set 2:

(1 comment)

Played with this since I was curious what the performance was, but it doesn't build on Mac either. The issues are:
- Mac spells .hidden as .private_extern.
- Functions accessed by C should be _.
- It's unhappy about all the label references (where it loads the table at the top). Clang complains with:
clang -cc1as: fatal error: error in backend: 32-bit absolute addressing is not supported in 64-bit mode

Linux shared library is unhappy about that last one too.
relocation R_X86_64_32S against `.data' can not be used when making a shared object; recompile with -fPIC
Probably we need to revise those to use rip-relative addressing? The first two we can fix with preprocessor (or even perlasm if we want it working on Windows).

Patch-set: 2